server/requirements.txt
fastapi>=0.115,<1.0
huggingface-hub>=0.25,<1.0
numpy>=2.0,<3.0
pillow>=10.4,<12.0
pydantic-settings>=2.4,<3.0
uvicorn[standard]>=0.30,<1.0

